% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_stats_table.R
\name{print.gt_data_table}
\alias{print.gt_data_table}
\title{Print an already summarised gtstats table}
\usage{
\method{print}{gt_data_table}(x, ...)
}
\arguments{
\item{x}{A \code{gt_data_table} object.}

\item{...}{Further arguments passed to \code{\link[=to_flextable]{to_flextable()}}.}
}
\value{
The input object, invisibly.
}
\description{
Print an already summarised gtstats table
}
