% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.gt_epi_table}
\alias{print.gt_epi_table}
\title{Print an outbreak and surveillance table}
\usage{
\method{print}{gt_epi_table}(x, ...)
}
\arguments{
\item{x}{A \code{gt_epi_table} object.}

\item{...}{Further arguments passed to \code{\link[=to_flextable]{to_flextable()}}.}
}
\value{
The input object, invisibly.
}
\description{
Print an outbreak and surveillance table
}
